RA/requirements_research.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
python-dotenv>=1.0.0
duckduckgo-search>=4.0.0
wikipedia>=1.4.0
PyQt6>=6.6.0
